+++ packages/access-definitions/src/Wrapping/Properties/ToManyRelationshipUpdatability.php
@@ -21,13 +21,13 @@ class ToManyRelationshipUpdatability extends AbstractRelationshipUpdatability
      * @param list<TCondition> $entityConditions
      * @param list<TCondition> $valueConditions
      * @param TransferableTypeInterface<TCondition, TSorting, TRelationship> $relationshipType
-     * @param null|callable(TEntity, iterable<TRelationship>): void $customWriteFunction
+     * @param null|callable(TEntity, iterable<TRelationship>): void $customWriteCallback
      */
     public function __construct(
         array $entityConditions,
         array $valueConditions,
         TransferableTypeInterface $relationshipType,
-        private $customWriteFunction
+        private readonly mixed $customWriteCallback
     ) {
         parent::__construct($entityConditions, $valueConditions, $relationshipType);
     }
@@ -35,8 +35,8 @@ class ToManyRelationshipUpdatability extends AbstractRelationshipUpdatability
     /**
      * @return null|callable(TEntity, iterable<TRelationship>): void
      */
-    public function getCustomWriteFunction(): ?callable
+    public function getCustomWriteCallback(): ?callable
     {
-        return $this->customWriteFunction;
+        return $this->customWriteCallback;
     }
 }